ft: filter by user and id

+++ Love_me_app/routers/product_review.py
@@ -42,19 +42,22 @@ def get_reviews( db:Session = Depends(get_db)):
     return reviews
 
 
-@router.delete('/del/{review_id}')
-def create_review(user:dict=Depends(get_current_user), db:Session = Depends(get_db)):
+@router.delete('/{review_id}')
+def delete_review(review_id, user:dict=Depends(get_current_user), db:Session = Depends(get_db)):
     current_user = user
     if current_user is not None:
-
-        review = db.query(ProductReviewModel).filter(id==id).first()
-        db.delete(review)
-        db.commit()
-        return {"Review  deleted"}
+        review = db.query(ProductReviewModel).filter(ProductReviewModel.user==current_user).filter(id==review_id).first()
+  
+        if review is not None:
+            
+            db.delete(review)
+            db.commit()
+            return {"Review  deleted"}
+        else:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail="Not Found")
     else:
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail="Unauthorized")
-    
-    
+
